Xong hóa đơn

+++ src/vn/ChuVoiCon/qlHangHoa/DLL/chi_tiet_hoa_don.java
@@ -2,17 +2,17 @@ package vn.ChuVoiCon.qlHangHoa.DLL;
 
 
 public class chi_tiet_hoa_don {
-	private long ma_hoa_don;
+	private int ma_hoa_don;
 	private String id_sp;
 	private chi_tiet_san_pham sp;
 	private double don_gia;
 	private int so_luong;
 	private hoa_don hoa_don;
 	
-	public long getMa_hoa_don() {
+	public int getMa_hoa_don() {
 		return ma_hoa_don;
 	}
-	public void setMa_hoa_don(long ma_hoa_don) {
+	public void setMa_hoa_don(int ma_hoa_don) {
 		this.ma_hoa_don = ma_hoa_don;
 	}
 	public String getId_sp() {
